Move component specs into their own folder
