+++ react-app/__tests__/components/TodoApp.react-test.js
@@ -1,12 +1,12 @@
 // __tests__/TodoApp.react-test.js
 
-jest.dontMock('../js/components/TodoApp.react.js');
+jest.dontMock('../../js/components/TodoApp.react.js');
 var React = require('react');
 var TestUtils = require('react-addons-test-utils');
-var TodoApp = require('../js/components/TodoApp.react.js');
-var TodoList = require('../js/components/TodoList.react.js');
-var LoginForm = require('../js/components/LoginForm.react.js');
-var AuthService = require('../js/services/AuthService.js');
+var TodoApp = require('../../js/components/TodoApp.react.js');
+var TodoList = require('../../js/components/TodoList.react.js');
+var LoginForm = require('../../js/components/LoginForm.react.js');
+var AuthService = require('../../js/services/AuthService.js');
 
 describe('TodoApp.react', () => {
   var promise = { then: () => { return promise; }, catch: () => { return promise; } };
